Download GOS file before building the documentation and use pooch caching in CI documentation build
