+++ exspy/conftest.py
@@ -17,7 +17,6 @@
 # along with eXSpy. If not, see <https://www.gnu.org/licenses/#GPL>.
 
 import importlib
-from time import sleep
 
 try:
     # Set traits toolkit to work in a headless system
@@ -37,10 +36,11 @@ import matplotlib.pyplot as plt
 import pytest
 import numpy as np
 import hyperspy.api as hs
-import pooch
 
 # Use matplotlib fixture to clean up figure, setup backend, etc.
 from matplotlib.testing.conftest import mpl_test_settings  # noqa: F401
+
+from exspy._misc import _download_GOS_files
 
 
 @pytest.fixture(autouse=True)
@@ -62,45 +62,6 @@ def setup_module(mod, pdb_cmdopt):
         dask.set_options(get=dask.local.get_sync)
 
 
-def _download_and_cache_GOS():
-    """Download GOS files for testing purposes."""
-
-    def retry_and_sleep(url, known_hash, retries=3, sleep_time=30):
-        for attempt in range(retries):
-            try:
-                return pooch.retrieve(
-                    url=url,
-                    known_hash=known_hash,
-                    downloader=pooch.HTTPDownloader(chunk_size=30000),
-                    progressbar=False,
-                )
-            except Exception as e:
-                if attempt < retries - 1:
-                    print(
-                        f"Download failed (attempt {attempt + 1}/{retries}). "
-                        f"Retrying in {sleep_time} seconds..."
-                    )
-                    sleep(sleep_time)
-                else:
-                    print("All download attempts failed.")
-                    raise e
-
-    print("Start download of GOS files")
-    retry_and_sleep(
-        url="https://zenodo.org/records/6599071/files/Segger_Guzzinati_Kohl_1.0.0.gos",
-        known_hash="md5:d65d5c23142532fde0a80e160ab51574",
-    )
-    retry_and_sleep(
-        url="https://zenodo.org/records/7645765/files/Segger_Guzzinati_Kohl_1.5.0.gosh",
-        known_hash="md5:7fee8891c147a4f769668403b54c529b",
-    )
-    retry_and_sleep(
-        url="https://zenodo.org/records/12800856/files/Dirac_GOS_compact.gosh",
-        known_hash="md5:01a855d3750d2c063955248358dbee8d",
-    )
-    print("finished download of GOS files")
-
-
 pytest_mpl_spec = importlib.util.find_spec("pytest_mpl")
 
 
@@ -112,11 +73,9 @@ if pytest_mpl_spec is None:
             "mpl_image_compare: dummy marker registration to allow running "
             "without the pytest-mpl plugin.",
         )
-        _download_and_cache_GOS()
-
+        _download_GOS_files()
 
 else:
 
     def pytest_configure(config):
-        _download_and_cache_GOS()
-        print("finish")
+        _download_GOS_files()
